requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart